+++ src/pypi_attestations/_impl.py
@@ -6,8 +6,9 @@ This module is NOT a public API, and is not considered stable.
 from __future__ import annotations
 
 import base64
+import json
 from enum import Enum
-from typing import TYPE_CHECKING, Annotated, Any, Literal, NewType, cast
+from typing import TYPE_CHECKING, Annotated, Any, Literal, NewType, cast, Optional, Union, get_args
 
 import sigstore.errors
 from annotated_types import MinLen  # noqa: TCH002
@@ -16,7 +17,7 @@ from cryptography.hazmat.primitives import serialization
 from packaging.utils import parse_sdist_filename, parse_wheel_filename
 from pyasn1.codec.der.decoder import decode as der_decode
 from pyasn1.type.char import UTF8String
-from pydantic import Base64Bytes, BaseModel, ConfigDict, Field, field_validator
+from pydantic import Base64Encoder, BaseModel, ConfigDict, EncodedBytes, Field, field_validator
 from pydantic.alias_generators import to_snake
 from pydantic_core import ValidationError
 from sigstore._utils import _sha256_streaming
@@ -25,15 +26,40 @@ from sigstore.dsse import Envelope as DsseEnvelope
 from sigstore.dsse import Error as DsseError
 from sigstore.models import Bundle, LogEntry
 from sigstore.sign import ExpiredCertificate, ExpiredIdentity
+from sigstore.verify import Verifier, policy
 from sigstore_protobuf_specs.io.intoto import Envelope as _Envelope
 from sigstore_protobuf_specs.io.intoto import Signature as _Signature
 
-if TYPE_CHECKING:
-    from pathlib import Path  # pragma: no cover
+if TYPE_CHECKING:  # pragma: no cover
+    from pathlib import Path
 
-    from sigstore.sign import Signer  # pragma: no cover
-    from sigstore.verify import Verifier  # pragma: no cover
-    from sigstore.verify.policy import VerificationPolicy  # pragma: no cover
+    from cryptography.x509 import Certificate
+    from sigstore.sign import Signer
+    from sigstore.verify.policy import VerificationPolicy
+
+
+class Base64EncoderSansNewline(Base64Encoder):
+    r"""A Base64Encoder that doesn't insert newlines when encoding.
+
+    Pydantic's Base64Bytes type inserts newlines b'\n' every 76 characters because they
+    use `base64.encodebytes()` instead of `base64.b64encode()`. Pydantic maintainers
+    have stated that they won't fix this, and that users should work around it by
+    defining their own Base64 type with a custom encoder.
+    See https://github.com/pydantic/pydantic/issues/9072 for more details.
+    """
+
+    @classmethod
+    def encode(cls, value: bytes) -> bytes:
+        """Encode bytes to base64."""
+        return base64.b64encode(value)
+
+    @classmethod
+    def decode(cls, value: bytes) -> bytes:
+        """Decode base64 bytes."""
+        return base64.b64decode(value, validate=True)
+
+
+Base64Bytes = Annotated[bytes, EncodedBytes(encoder=Base64EncoderSansNewline)]
 
 
 class Distribution(BaseModel):
@@ -123,6 +149,15 @@ class Attestation(BaseModel):
     The enveloped attestation statement and signature.
     """
 
+    @property
+    def statement(self) -> dict[str, Any]:
+        """Return the statement within this attestation's envelope.
+
+        The value returned here is a dictionary, in the shape of an
+        in-toto statement.
+        """
+        return json.loads(self.envelope.statement)  # type: ignore[no-any-return]
+
     @classmethod
     def sign(cls, signer: Signer, dist: Distribution) -> Attestation:
         """Create an envelope, with signature, from the given Python distribution.
@@ -183,14 +218,37 @@ class Attestation(BaseModel):
 
     def verify(
         self,
-        verifier: Verifier,
-        policy: VerificationPolicy,
+        identity: VerificationPolicy | Publisher,
         dist: Distribution,
-    ) -> tuple[str, dict[str, Any] | None]:
+        *,
+        staging: bool = False,
+    ) -> tuple[str, Optional[dict[str, Any]]]:
         """Verify against an existing Python distribution.
+
+        The `identity` can be an object confirming to
+        `sigstore.policy.VerificationPolicy` or a `Publisher`, which will be
+        transformed into an appropriate verification policy.
+
+        By default, Sigstore's production verifier will be used. The
+        `staging` parameter can be toggled to enable the staging verifier
+        instead.
 
         On failure, raises an appropriate subclass of `AttestationError`.
         """
+        # NOTE: Can't do `isinstance` with `Publisher` since it's
+        # a `_GenericAlias`; instead we punch through to the inner
+        # `_Publisher` union.
+        # Use of typing.get_args is needed for Python < 3.10
+        if isinstance(identity, get_args(_Publisher)):
+            policy = identity._as_policy()  # noqa: SLF001
+        else:
+            policy = identity
+
+        if staging:
+            verifier = Verifier.staging()
+        else:
+            verifier = Verifier.production()
+
         bundle = self.to_bundle()
         try:
             type_, payload = verifier.verify_dsse(bundle, policy)
@@ -351,12 +409,15 @@ def _ultranormalize_dist_filename(dist: str) -> str:
             parts = "-".join([name, str(ver), impl_tag, abi_tag, platform_tag])
 
         return f"{parts}.whl"
-
-    elif dist.endswith(".tar.gz"):
+    elif dist.endswith((".tar.gz", ".zip")):
         # `parse_sdist_filename` raises a supertype of ValueError on failure.
         name, ver = parse_sdist_filename(dist)
         name = name.replace("-", "_")
-        return f"{name}-{ver}.tar.gz"
+
+        if dist.endswith(".tar.gz"):
+            return f"{name}-{ver}.tar.gz"
+        else:
+            return f"{name}-{ver}.zip"
     else:
         raise ValueError(f"unknown distribution format: {dist}")
 
@@ -365,7 +426,83 @@ class _PublisherBase(BaseModel):
     model_config = ConfigDict(alias_generator=to_snake)
 
     kind: str
-    claims: dict[str, Any] | None = None
+    claims: Optional[dict[str, Any]] = None
+
+    def _as_policy(self) -> VerificationPolicy:
+        """Return an appropriate `sigstore.policy.VerificationPolicy` for this publisher."""
+        raise NotImplementedError  # pragma: no cover
+
+
+class _GitHubTrustedPublisherPolicy:
+    """A custom sigstore-python policy for verifying against a GitHub-based Trusted Publisher."""
+
+    def __init__(self, repository: str, workflow: str) -> None:
+        self._repository = repository
+        self._workflow = workflow
+        # This policy must also satisfy some baseline underlying policies:
+        # the issuer must be GitHub Actions, and the repo must be the one
+        # we expect.
+        self._subpolicy = policy.AllOf(
+            [
+                policy.OIDCIssuerV2("https://token.actions.githubusercontent.com"),
+                policy.OIDCSourceRepositoryURI(f"https://github.com/{self._repository}"),
+            ]
+        )
+
+    @classmethod
+    def _der_decode_utf8string(cls, der: bytes) -> str:
+        """Decode a DER-encoded UTF8String."""
+        return der_decode(der, UTF8String)[0].decode()  # type: ignore[no-any-return]
+
+    def verify(self, cert: Certificate) -> None:
+        """Verify the certificate against the Trusted Publisher identity."""
+        self._subpolicy.verify(cert)
+
+        # This process has a few annoying steps, since a Trusted Publisher
+        # isn't aware of the commit or ref it runs on, while Sigstore's
+        # leaf certificate claims (like GitHub Actions' OIDC claims) only
+        # ever encode the workflow filename (which we need to check) next
+        # to the ref/sha (which we can't check).
+        #
+        # To get around this, we:
+        # (1) extract the `Build Config URI` extension;
+        # (2) extract the `Source Repository Digest` and
+        #     `Source Repository Ref` extensions;
+        # (3) build the *expected* URI with the user-controlled
+        #     Trusted Publisher identity *with* (2)
+        # (4) compare (1) with (3)
+
+        # (1) Extract the build config URI, which looks like this:
+        #     https://github.com/OWNER/REPO/.github/workflows/WORKFLOW@REF
+        #  where OWNER/REPO and WORKFLOW are controlled by the TP identity,
+        #  and REF is controlled by the certificate's own claims.
+        build_config_uri = cert.extensions.get_extension_for_oid(policy._OIDC_BUILD_CONFIG_URI_OID)  # noqa: SLF001
+        raw_build_config_uri = self._der_decode_utf8string(build_config_uri.value.public_bytes())
+
+        # (2) Extract the source repo digest and ref.
+        source_repo_digest = cert.extensions.get_extension_for_oid(
+            policy._OIDC_SOURCE_REPOSITORY_DIGEST_OID  # noqa: SLF001
+        )
+        sha = self._der_decode_utf8string(source_repo_digest.value.public_bytes())
+
+        source_repo_ref = cert.extensions.get_extension_for_oid(
+            policy._OIDC_SOURCE_REPOSITORY_REF_OID  # noqa: SLF001
+        )
+        ref = self._der_decode_utf8string(source_repo_ref.value.public_bytes())
+
+        # (3)-(4): Build the expected URIs and compare them
+        for suffix in [sha, ref]:
+            expected = (
+                f"https://github.com/{self._repository}/.github/workflows/{self._workflow}@{suffix}"
+            )
+            if raw_build_config_uri == expected:
+                return
+
+        # If none of the expected URIs matched, the policy fails.
+        raise sigstore.errors.VerificationError(
+            f"Certificate's Build Config URI ({build_config_uri}) does not match expected "
+            f"Trusted Publisher ({self._workflow} @ {self._repository})"
+        )
 
 
 class GitHubPublisher(_PublisherBase):
@@ -385,11 +522,14 @@ class GitHubPublisher(_PublisherBase):
     action.
     """
 
-    environment: str | None = None
+    environment: Optional[str] = None
     """
     The optional name GitHub Actions environment that the publishing
     action was performed from.
     """
+
+    def _as_policy(self) -> VerificationPolicy:
+        return _GitHubTrustedPublisherPolicy(self.repository, self.workflow)
 
 
 class GitLabPublisher(_PublisherBase):
@@ -404,13 +544,34 @@ class GitLabPublisher(_PublisherBase):
     `bar` owned by group `foo` and subgroup `baz`.
     """
 
-    environment: str | None = None
+    environment: Optional[str] = None
     """
     The optional environment that the publishing action was performed from.
     """
 
+    def _as_policy(self) -> VerificationPolicy:
+        policies: list[VerificationPolicy] = [
+            policy.OIDCIssuerV2("https://gitlab.com"),
+            policy.OIDCSourceRepositoryURI(f"https://gitlab.com/{self.repository}"),
+        ]
 
-Publisher = Annotated[GitHubPublisher | GitLabPublisher, Field(discriminator="kind")]
+        if not self.claims:
+            raise VerificationError("refusing to build a policy without claims")
+
+        if ref := self.claims.get("ref"):
+            policies.append(
+                policy.OIDCBuildConfigURI(
+                    f"https://gitlab.com/{self.repository}//.gitlab-ci.yml@{ref}"
+                )
+            )
+        else:
+            raise VerificationError("refusing to build a policy without a ref claim")
+
+        return policy.AllOf(policies)
+
+
+_Publisher = Union[GitHubPublisher, GitLabPublisher]
+Publisher = Annotated[_Publisher, Field(discriminator="kind")]
 
 
 class AttestationBundle(BaseModel):