+++ CHANGELOG.md
@@ -13,6 +13,58 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   signing certificate extensions
   ([#XX](https://github.com/trailofbits/pypi-attestations/pull/XX))
 
+## [0.0.16]
+
+### Added
+
+- `Attestation.statement` has been added as a convenience API for accessing
+  the attestation's enveloped statement as a dictionary
+
+## [0.0.15]
+
+This is a corrective release for [0.0.14].
+
+## [0.0.14]
+
+### Fixed
+
+- The `Distribution` API now handles ZIP source distributions
+  (those ending with `.zip`) instead of rejecting them as invalid
+  ([#68](https://github.com/trailofbits/pypi-attestations/pull/68))
+
+## [0.0.13]
+
+### Changed
+
+- The minimum Python version required has been brought back to `3.9`
+  ([#64](https://github.com/trailofbits/pypi-attestations/pull/64)).
+
+- The `Attestation.verify(...)` API has been changed to remove the `Verifier`
+  argument in favor of an optional `staging: bool` kwarg to select the
+  Sigstore instance
+  ([#62](https://github.com/trailofbits/pypi-attestations/pull/62))
+
+- The `Attestation.verify(...)` API has been changed to accept both `Publisher`
+  and `VerificationPolicy` objects as a policy. The publisher object is internally
+  converted to an appropriate verification policy.
+
+### Fixed
+
+- `python -m pypi_attestations verify` now handles inputs like `dist/*`
+  gracefully, by pre-filtering any attestation paths from the inputs.
+
+- `python -m pypi_attestations verify` now exits with a non-zero exit code
+  if the verification step fails
+  ([#57](https://github.com/trailofbits/pypi-attestations/pull/57))
+
+## [0.0.12]
+
+### Fixed
+
+- Base64-encoded bytes inside Attestation objects contained newline characters
+  every 76 characters due to a bug in Pydantic's Base64Bytes type. Those
+  newlines were also (incorrectly) ignored by Pydantic during decoding
+  ([#48](https://github.com/trailofbits/pypi-attestations/pull/48)).
 
 ## [0.0.11]
 
@@ -128,7 +180,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation
 
-[Unreleased]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.11...HEAD
+[Unreleased]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.16...HEAD
+[0.0.16]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.15...v0.0.15
+[0.0.15]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.14...v0.0.15
+[0.0.14]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.13...v0.0.14
+[0.0.13]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.12...v0.0.13
+[0.0.12]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.11...v0.0.12
 [0.0.11]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.10...v0.0.11
 [0.0.10]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.8...v0.0.9